Added observations.tex to the research folder, which documents a lot of the (relatively) interesting things I've come across during my research.
Added new_MultiVarExtMatrixT(), a function for creating a GenericMatrixT specifically for MultiVarExtTs. Basically, it prevents the user from having to type a lot of functions.

Changed the behaviour of set_GenericMatrixT(). It now requires that the matrix be initalised first. This is so that the same matrix can be set multiple times without memory leaking.

Added get_MultiVarExtT_coefficient(), a way to get a MultiVarExtT's coefficient at a specific location.